LayoutTests/storage/indexeddb/modern/createobjectstore-failures.html

This test exercises the obvious ways that IDBDatabase.createObjectStore can fail.
<script>

if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
}

function done()
{
    alert("Done");
    if (window.testRunner)
        testRunner.notifyDone();
}

function dumpObjectStores(database) {
    var list = database.objectStoreNames;
    alert("Object store names:");
    for (var i = 0; i < list.length; ++i) { 
        alert(list[i]);
    }
}

var createRequest = window.indexedDB.open("CreateObjectStoreFailuresTestDatabase", 1);
var database;
var versionTransaction;

createRequest.onupgradeneeded = function(event) {
    alert("Initial upgrade needed: Old version - " + event.oldVersion + " New version - " + event.newVersion);

    versionTransaction = createRequest.transaction;
    database = event.target.result;

    try {
        var objectStore = database.createObjectStore('TestObjectStore', { autoIncrement: true , keyPath: "" });
    } catch (e) {
        alert("Failed to create object store with both autoincrement and an empty keypath: " + e);
        dumpObjectStores(database);
    }

    try {
        var objectStore = database.createObjectStore('TestObjectStore', { autoIncrement: true , keyPath: ['foo'] });
    } catch (e) {
        alert("Failed to create object store with both autoincrement and a sequence keypath: " + e);
        dumpObjectStores(database);
    }
    
    try {
        var objectStore = database.createObjectStore('TestObjectStore', { keyPath: "'foo bar'"});
    } catch (e) {
        alert("Failed to create object store with invalid keyPath: " + e);
        dumpObjectStores(database);
    }

    database.createObjectStore("TestObjectStore1");
    alert("Actually created an object store");
    dumpObjectStores(database);
    
    try {
        database.createObjectStore("TestObjectStore1");
    } catch(e) {
        alert("Failed to create TestObjectStore a second time: " + e);
        dumpObjectStores(database);
    }

    versionTransaction.onabort = function(event) {
        alert("Initial upgrade versionchange unexpected abort");
        done();
    }

    versionTransaction.oncomplete = function(event) {
        alert("Initial upgrade versionchange transaction complete");
        dumpObjectStores(database);

        try {
            database.createObjectStore("TestObjectStore2");
        } catch(e) {
            alert("Failed to create object store while there is no version change transaction: " + e);
            dumpObjectStores(database);
        }
    
        setTimeout(finishUp, 0);
    }

    versionTransaction.onerror = function(event) {
        alert("Initial upgrade versionchange transaction unexpected error" + event);
        done();
    }
}

function finishUp()
{
    try {
        database.createObjectStore("TestObjectStore99");
    } catch(e) {
        alert("Failed to create object store outside of onupgradeneeded: " + e);
        dumpObjectStores(database);
    }
    
    done();
}

</script>
