LayoutTests/storage/indexeddb/modern/createobjectstore-failures-expected.txt

ALERT: Initial upgrade needed: Old version - 0 New version - 1
ALERT: Failed to create object store with both autoincrement and an empty keypath: Error: InvalidAccessError: DOM IDBDatabase Exception 15
ALERT: Object store names:
ALERT: Failed to create object store with both autoincrement and a sequence keypath: Error: InvalidAccessError: DOM IDBDatabase Exception 15
ALERT: Object store names:
ALERT: Failed to create object store with invalid keyPath: Error: SyntaxError: DOM IDBDatabase Exception 12
ALERT: Object store names:
ALERT: Actually created an object store
ALERT: Object store names:
ALERT: TestObjectStore1
ALERT: Failed to create TestObjectStore a second time: Error: ConstraintError: DOM IDBDatabase Exception 0
ALERT: Object store names:
ALERT: TestObjectStore1
ALERT: Initial upgrade versionchange transaction complete
ALERT: Object store names:
ALERT: TestObjectStore1
ALERT: Failed to create object store while there is no version change transaction: Error: InvalidStateError: DOM IDBDatabase Exception 11
ALERT: Object store names:
ALERT: TestObjectStore1
ALERT: Failed to create object store outside of onupgradeneeded: Error: InvalidStateError: DOM IDBDatabase Exception 11
ALERT: Object store names:
ALERT: TestObjectStore1
ALERT: Done
This test exercises the obvious ways that IDBDatabase.createObjectStore can fail.
